Change floats to doubles, add P calculation
Changelog:
    - Change all floats to doubles
    - Add different elastic material models
        - Linear elasticity
        - St. Venant Kirchhoff
        - Corotated Linear
        - Neo-Hookean
    - Add correct tetrahedron vertex mass calculation
    - Fix for small floating point errors causing infinite growing elastic force
    - Add ground collision detection
    - Add system to simulate multiple time steps per written frame
    - Add timing
    - Add more status messages
    - Add force attribute for output .bgeo files
    - Add function to add uniform force
    - Add function to reset all forces

TODO:
    - Simulation currently causes the particles to disappear when colliding up with
      the ground

+++ Projects/main/tastyio.h
@@ -8,8 +8,8 @@
 namespace TastyIO {
 
 void readOBJ(const std::string &objFilename, tetgenio &out);
-void parseNodeFile(const std::string &inputNodeFile, Particles<float, 3> &ps);
-void parseEleFile(const std::string &inputEleFile, std::vector<Tetra<float, 3>> &tets, float k, Particles<float, 3> &p);
+void parseNodeFile(const std::string &inputNodeFile, Particles<double, 3> &ps);
+void parseEleFile(const std::string &inputEleFile, std::vector<Tetra<double, 3>> &tets, double k, Particles<double, 3> &p);
 
 }
